refactor(tui): extract bot avatar rendering to render_bots.rs (#358)
Move ~350 lines of self-contained generative bot rendering code from
render.rs to a new render_bots.rs module:
- BotBorder, BotEyes, BotMouth, WinkStyle enums
- BotParts struct and bot_from_seed() (splitmix64 deterministic assembly)
- slot_hash(), bot_color(), border_chars(), eye_char(), mouth_str(),
  mouth_talk_str(), bot_row() (per-row rendering with animation)
- welcome_splash() (the only pub(super) export — re-exported through render.rs)
- Animation constants (BLINK_INTERVAL, SLOT_FRAMES, WINK_CHANCE, etc.)
- 8 welcome_splash tests

render.rs drops from 1235 to ~850 lines. No behavior changes.

Closes #358
